requirements.txt
cryptography
argon2-cffi
PySide6
pyinstaller
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
